Rm log.hpp

+++ src/d3c/scene/scene_portal.cpp
@@ -1,5 +1,4 @@
 #include "scene_portal.hpp"
-#include "../misc/log.hpp"
 #include "../../engine/screen.hh"
 #include "../parse_common.hh"
 
@@ -173,7 +172,7 @@ void Portal_portal::read_from_file(std::ifstream &file) {
 int Portal_portal::check_visibility(const camera *cam) {
   bool all_front = true, all_back = true;
   // check front plane
-  for (int i = 0; i < m_points.size(); ++i)
+  for (size_t i = 0; i < m_points.size(); ++i)
     if (!renderer->get_frustum_plane(frustum_plane_front).is_in_front(m_points[i]))
       all_front = false;
     else
@@ -185,7 +184,7 @@ int Portal_portal::check_visibility(const camera *cam) {
   // check other planes
   for (int j = frustum_plane_front + 1; j < frustum_plane_back; ++j) {
     bool all_back = true;
-    for (int i = 0; i < m_points.size(); ++i)
+    for (size_t i = 0; i < m_points.size(); ++i)
       if (renderer->get_frustum_plane(j).is_in_front(m_points[i])) {
         all_back = false;
         break;
